requirements.txt
speechrecognition
pyttsx3
PyAudio
google.generativeai
transformers
setuptools
python-dotenv